[Autocomplete] Refactor downshift demos to minimize changes
