+++ docs/src/pages/demos/autocomplete/IntegrationDownshift.js
@@ -2,11 +2,93 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import deburr from 'lodash/deburr';
 import Downshift from 'downshift';
-import { withStyles } from '@material-ui/core/styles';
+import { makeStyles } from '@material-ui/core/styles';
+import TextField from '@material-ui/core/TextField';
 import Popper from '@material-ui/core/Popper';
 import Paper from '@material-ui/core/Paper';
+import MenuItem from '@material-ui/core/MenuItem';
 import Chip from '@material-ui/core/Chip';
-import { suggestions, styles, renderInput, renderSuggestion } from './IntegrationDownshift.common';
+
+export const suggestions = [
+  { label: 'Afghanistan' },
+  { label: 'Aland Islands' },
+  { label: 'Albania' },
+  { label: 'Algeria' },
+  { label: 'American Samoa' },
+  { label: 'Andorra' },
+  { label: 'Angola' },
+  { label: 'Anguilla' },
+  { label: 'Antarctica' },
+  { label: 'Antigua and Barbuda' },
+  { label: 'Argentina' },
+  { label: 'Armenia' },
+  { label: 'Aruba' },
+  { label: 'Australia' },
+  { label: 'Austria' },
+  { label: 'Azerbaijan' },
+  { label: 'Bahamas' },
+  { label: 'Bahrain' },
+  { label: 'Bangladesh' },
+  { label: 'Barbados' },
+  { label: 'Belarus' },
+  { label: 'Belgium' },
+  { label: 'Belize' },
+  { label: 'Benin' },
+  { label: 'Bermuda' },
+  { label: 'Bhutan' },
+  { label: 'Bolivia, Plurinational State of' },
+  { label: 'Bonaire, Sint Eustatius and Saba' },
+  { label: 'Bosnia and Herzegovina' },
+  { label: 'Botswana' },
+  { label: 'Bouvet Island' },
+  { label: 'Brazil' },
+  { label: 'British Indian Ocean Territory' },
+  { label: 'Brunei Darussalam' },
+];
+
+export function renderInput(inputProps) {
+  const { InputProps, classes, ref, ...other } = inputProps;
+
+  return (
+    <TextField
+      InputProps={{
+        inputRef: ref,
+        classes: {
+          root: classes.inputRoot,
+          input: classes.inputInput,
+        },
+        ...InputProps,
+      }}
+      {...other}
+    />
+  );
+}
+
+export function renderSuggestion({ suggestion, index, itemProps, highlightedIndex, selectedItem }) {
+  const isHighlighted = highlightedIndex === index;
+  const isSelected = (selectedItem || '').indexOf(suggestion.label) > -1;
+
+  return (
+    <MenuItem
+      {...itemProps}
+      key={suggestion.label}
+      selected={isHighlighted}
+      component="div"
+      style={{
+        fontWeight: isSelected ? 500 : 400,
+      }}
+    >
+      {suggestion.label}
+    </MenuItem>
+  );
+}
+renderSuggestion.propTypes = {
+  highlightedIndex: PropTypes.number,
+  index: PropTypes.number,
+  itemProps: PropTypes.object,
+  selectedItem: PropTypes.string,
+  suggestion: PropTypes.shape({ label: PropTypes.string }).isRequired,
+};
 
 function getSuggestions(value) {
   const inputValue = deburr(value.trim()).toLowerCase();
@@ -27,113 +109,129 @@ function getSuggestions(value) {
       });
 }
 
-class DownshiftMultiple extends React.Component {
-  state = {
-    inputValue: '',
-    selectedItem: [],
-  };
+function DownshiftMultiple(props) {
+  const { classes } = props;
+  const [inputValue, setInputValue] = React.useState('');
+  const [selectedItem, setSelectedItem] = React.useState([]);
 
-  handleKeyDown = event => {
-    const { inputValue, selectedItem } = this.state;
+  function handleKeyDown(event) {
     if (selectedItem.length && !inputValue.length && event.key === 'Backspace') {
-      this.setState({
-        selectedItem: selectedItem.slice(0, selectedItem.length - 1),
-      });
+      setSelectedItem(selectedItem.slice(0, selectedItem.length - 1));
     }
-  };
-
-  handleInputChange = event => {
-    this.setState({ inputValue: event.target.value });
-  };
-
-  handleChange = item => {
-    let { selectedItem } = this.state;
-
-    if (selectedItem.indexOf(item) === -1) {
-      selectedItem = [...selectedItem, item];
-    }
-
-    this.setState({
-      inputValue: '',
-      selectedItem,
-    });
-  };
-
-  handleDelete = item => () => {
-    this.setState(state => {
-      const selectedItem = [...state.selectedItem];
-      selectedItem.splice(selectedItem.indexOf(item), 1);
-      return { selectedItem };
-    });
-  };
-
-  render() {
-    const { classes } = this.props;
-    const { inputValue, selectedItem } = this.state;
-
-    return (
-      <Downshift
-        id="downshift-multiple"
-        inputValue={inputValue}
-        onChange={this.handleChange}
-        selectedItem={selectedItem}
-      >
-        {({
-          getInputProps,
-          getItemProps,
-          isOpen,
-          inputValue: inputValue2,
-          selectedItem: selectedItem2,
-          highlightedIndex,
-        }) => (
-          <div className={classes.container}>
-            {renderInput({
-              fullWidth: true,
-              classes,
-              InputProps: getInputProps({
-                startAdornment: selectedItem.map(item => (
-                  <Chip
-                    key={item}
-                    tabIndex={-1}
-                    label={item}
-                    className={classes.chip}
-                    onDelete={this.handleDelete(item)}
-                  />
-                )),
-                onChange: this.handleInputChange,
-                onKeyDown: this.handleKeyDown,
-                placeholder: 'Select multiple countries',
-              }),
-              label: 'Label',
-            })}
-            {isOpen ? (
-              <Paper className={classes.paper} square>
-                {getSuggestions(inputValue2).map((suggestion, index) =>
-                  renderSuggestion({
-                    suggestion,
-                    index,
-                    itemProps: getItemProps({ item: suggestion.label }),
-                    highlightedIndex,
-                    selectedItem: selectedItem2,
-                  }),
-                )}
-              </Paper>
-            ) : null}
-          </div>
-        )}
-      </Downshift>
-    );
   }
+
+  function handleInputChange(event) {
+    setInputValue(event.target.value);
+  }
+
+  function handleChange(item) {
+    let newSelectedItem = [...selectedItem];
+    if (newSelectedItem.indexOf(item) === -1) {
+      newSelectedItem = [...newSelectedItem, item];
+    }
+    setInputValue('');
+    setSelectedItem(newSelectedItem);
+  }
+
+  const handleDelete = item => () => {
+    const newSelectedItem = [...selectedItem];
+    newSelectedItem.splice(newSelectedItem.indexOf(item), 1);
+    setSelectedItem(newSelectedItem);
+  };
+
+  return (
+    <Downshift
+      id="downshift-multiple"
+      inputValue={inputValue}
+      onChange={handleChange}
+      selectedItem={selectedItem}
+    >
+      {({
+        getInputProps,
+        getItemProps,
+        isOpen,
+        inputValue: inputValue2,
+        selectedItem: selectedItem2,
+        highlightedIndex,
+      }) => (
+        <div className={classes.container}>
+          {renderInput({
+            fullWidth: true,
+            classes,
+            InputProps: getInputProps({
+              startAdornment: selectedItem.map(item => (
+                <Chip
+                  key={item}
+                  tabIndex={-1}
+                  label={item}
+                  className={classes.chip}
+                  onDelete={handleDelete(item)}
+                />
+              )),
+              onChange: handleInputChange,
+              onKeyDown: handleKeyDown,
+              placeholder: 'Select multiple countries',
+            }),
+            label: 'Label',
+          })}
+          {isOpen ? (
+            <Paper className={classes.paper} square>
+              {getSuggestions(inputValue2).map((suggestion, index) =>
+                renderSuggestion({
+                  suggestion,
+                  index,
+                  itemProps: getItemProps({ item: suggestion.label }),
+                  highlightedIndex,
+                  selectedItem: selectedItem2,
+                }),
+              )}
+            </Paper>
+          ) : null}
+        </div>
+      )}
+    </Downshift>
+  );
 }
 
 DownshiftMultiple.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
+export const useStyles = makeStyles(theme => ({
+  root: {
+    flexGrow: 1,
+    height: 250,
+  },
+  container: {
+    flexGrow: 1,
+    position: 'relative',
+  },
+  paper: {
+    position: 'absolute',
+    zIndex: 1,
+    marginTop: theme.spacing(1),
+    left: 0,
+    right: 0,
+  },
+  chip: {
+    margin: theme.spacing(0.5, 0.25),
+  },
+  inputRoot: {
+    flexWrap: 'wrap',
+  },
+  inputInput: {
+    width: 'auto',
+    flexGrow: 1,
+  },
+  divider: {
+    height: theme.spacing(2),
+  },
+}));
+
 let popperNode;
 
-function IntegrationDownshift(props) {
-  const { classes } = props;
+function IntegrationDownshift() {
+  const classes = useStyles();
 
   return (
     <div className={classes.root}>
@@ -222,8 +320,4 @@ function IntegrationDownshift(props) {
   );
 }
 
-IntegrationDownshift.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(styles)(IntegrationDownshift);
+export default IntegrationDownshift;
